# LibCompress: Decode non-self-referencing back-references in one shot